backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
sentence-transformers
transformers
torch
numpy
tqdm